Unswap for better ordering
