+++ tests/examples/diet_problem/results/problem-copt-machine.lp
@@ -13,10 +13,10 @@ Subject To
  R06: C07 - 60 ONE <= 0
  R07: C08 - 6.8 ONE <= 0
  R08: C09 - 80 ONE <= 0
- R09: - 410 C01 - 420 C02 - 560 C03 - 380 C04 - 320 C05 - 320 C06 - 320 C07 
-   - 100 C08 - 330 C09 + 1800 ONE <= 0
- R10: - 24 C01 - 32 C02 - 20 C03 - 4 C04 - 12 C05 - 15 C06 - 31 C07 - 8 C08 
-   - 8 C09 + 91 ONE <= 0
+ R09: 1800 ONE - 410 C01 - 420 C02 - 560 C03 - 380 C04 - 320 C05 - 320 C06 
+   - 320 C07 - 100 C08 - 330 C09 <= 0
+ R10: 91 ONE - 24 C01 - 32 C02 - 20 C03 - 4 C04 - 12 C05 - 15 C06 - 31 C07 
+   - 8 C08 - 8 C09 <= 0
  R11: 410 C01 + 420 C02 + 560 C03 + 380 C04 + 320 C05 + 320 C06 + 320 C07 
    + 100 C08 + 330 C09 - 2200 ONE <= 0
  R12: 26 C01 + 10 C02 + 32 C03 + 19 C04 + 10 C05 + 12 C06 + 12 C07
